Move Reader from serializer.cc's anon namespace to within impl namespace
This should be bug-for-bug compatible. No changes were made to the moved
methods.
