+++ Firestore/core/src/firebase/firestore/remote/impl/CMakeLists.txt
@@ -13,17 +13,13 @@
 # limitations under the License.
 
 cc_library(
-  firebase_firestore_remote
+  firebase_firestore_remote_impl
   SOURCES
-    datastore.h
-    datastore.cc
-    serializer.h
-    serializer.cc
+    tag.h
+    reader.h
+    reader.cc
   DEPENDS
-    firebase_firestore_model
-    firebase_firestore_protos_nanopb
-    firebase_firestore_remote_impl
     firebase_firestore_util
-    grpc::grpc
+    firebase_firestore_protos_nanopb
     nanopb
 )
